add caller verification for ipc with messenger

+++ src/frontends/android/app/src/main/java/org/strongswan/android/ipc/VpnProfileCrudMessengerService.java
@@ -9,7 +9,9 @@ import android.app.Service;
 import android.content.Intent;
 import android.os.*;
 import android.util.Log;
+import com.fancyfon.mobile.android.verification.CallerVerificator;
 import com.fancyfon.strongswan.api.R;
+import org.strongswan.android.ipc.verification.StrongswanCallerVerificator;
 
 import java.util.List;
 
@@ -20,26 +22,37 @@ public class VpnProfileCrudMessengerService extends Service {
 
     private static final String TAG = "VpnProfileService";
     private VpnProfileCrud vpnProfileCrud;
-
+    private CallerVerificator callerVerificator;
+    private int uid;
     Messenger messenger = new Messenger(new Handler() {
-
+        @Override
+        public boolean sendMessageAtTime(Message msg, long uptimeMillis) {
+            uid = Binder.getCallingUid();
+            return super.sendMessageAtTime(msg, uptimeMillis);
+        }
         @Override
         public void handleMessage(Message msg) {
-            if (msg.what == getInteger(R.integer.vpn_profile_create_message)) {
-                create(msg);
-            } else if (msg.what == getInteger(R.integer.vpn_profile_read_message)) {
-                read(msg);
-            } else if (msg.what == getInteger(R.integer.vpn_profile_read_all_message)) {
-                readAll(msg);
-            } else if (msg.what == getInteger(R.integer.vpn_profile_update_message)) {
-                update(msg);
-            } else if (msg.what == getInteger(R.integer.vpn_profile_delete_message)) {
-                delete(msg);
-            } else if (msg.what == getInteger(R.integer.vpn_profile_delete_all_message)) {
-                deleteAll(msg);
-            } else {
-                Log.w(TAG, "Unknown message: " + msg);
-                super.handleMessage(msg);
+            if ( callerVerificator.isCallerPermitted(uid)) {
+                if (msg.what == getInteger(R.integer.vpn_profile_create_message)) {
+                    create(msg);
+                } else if (msg.what == getInteger(R.integer.vpn_profile_read_message)) {
+                    read(msg);
+                } else if (msg.what == getInteger(R.integer.vpn_profile_read_all_message)) {
+                    readAll(msg);
+                } else if (msg.what == getInteger(R.integer.vpn_profile_update_message)) {
+                    update(msg);
+                } else if (msg.what == getInteger(R.integer.vpn_profile_delete_message)) {
+                    delete(msg);
+                } else if (msg.what == getInteger(R.integer.vpn_profile_delete_all_message)) {
+                    deleteAll(msg);
+                } else {
+                    Log.w(TAG, "Unknown message: " + msg);
+                    super.handleMessage(msg);
+                }
+            }else{
+                Log.e(TAG, "Cannot verify caller  uid " +uid  );
+                reply(msg,false);
+
             }
         }
 
@@ -117,13 +130,14 @@ public class VpnProfileCrudMessengerService extends Service {
 
     @Override
     public IBinder onBind(Intent intent) {
-        return messenger.getBinder();
+           return messenger.getBinder();
     }
 
     @Override
     public void onCreate() {
         super.onCreate();
         vpnProfileCrud = new VpnProfileCrud(this);
+        callerVerificator=new StrongswanCallerVerificator(getApplicationContext());
     }
 
     @Override
